Move Particle Filter agent to 'wifi' folder
