+++ reinforced_lib/exts/__init__.py
@@ -1,1 +1,2 @@
+from reinforced_lib.exts.base_ext import BaseExt
 from reinforced_lib.exts.ieee_802_11_ax import IEEE_802_11_ax
